clicking on a follower from the followers list result in loading that users profile
