+++ crt-app/src/App.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Switch, Route } from 'react-router-dom';
+import { connect } from 'react-redux';
 
 import './App.css';
 
@@ -9,28 +10,50 @@ import LandingPageMain from './components/landing-page/LP-main';
 import ListOfUsersAndSearch from './components/ListOfUsers/ListOfUsersAndSearch';
 import ProfilePage from "./components/ProfilePage/ProfilePage"
 
-function App() {
-  return (
-    <div className="App">
-      <NavBar/>
+class App extends React.Component {
 
-      <Switch>
-        <Route exact path='/' >
-          <LandingPageMain/>
-        </Route>
 
-        <Route path='/listOfUsers' >
-          <ListOfUsersAndSearch/>
-        </Route>
-        
-        <Route>
-          <ProfilePage/>
-        </Route>
-      </Switch>
-
-      <SubscribeFooter/>
-    </div>
-  );
+  render() {
+    return (
+      <div className="App">
+        <NavBar/>
+  
+        <Switch>
+          <Route exact path='/' >
+            <LandingPageMain/>
+          </Route>
+  
+          <Route path='/listOfUsers' >
+            <ListOfUsersAndSearch/>
+          </Route>
+          
+          {
+            this.props.user
+            ?
+            <Route path={`/profilePage/${this.props.user.login}`} >
+              <ProfilePage/>
+            </Route>
+            :
+            <Route path={`/profilePage`} >
+              <ProfilePage/>
+            </Route>
+          }
+          
+        </Switch>
+  
+        <SubscribeFooter/>
+      </div>
+    )
+  }
+  
 }
 
-export default App;
+const mapStateToProps = (state) => ({
+  user: state.users.user
+})
+
+
+const mapDispatchToProps = {
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(App);
